Rename OptimisticObjectCache to DepTrackingCache.
Although `OptimisticObjectCache` uses the `optimism` npm package, calling
it "optimistic" was actively misleading, since it has very little to do
with optimistic updates, which is what "optimistic" refers to almost
everywhere else in this codebase. In fact, if any `NormalizedCache`
subclass deserves to be called "optimistic," it's the `RecordingCache`,
which exists precisely for the purpose of recording optimistic updates
that should not be written (yet) to the official cache.