Mention FetchPolicy/WatchQueryFetchPolicy split in CHANGELOG.md.
Because they happen at build time, type errors are in a fundamentally
different category than runtime errors. We reserve the right to adjust our
types from time to time, and we rely on TypeScript to communicate those
changes to developers.

Type errors may require some action (hence this note in CHANGELOG.md) but
they communicate worthwhile information and are always worth fixing.

If you're not ready to deal with type-level changes, please use a
package-lock.json or yarn.lock file to pin the exact versions of your
dependencies.

Should help with #4847, along with the recent release of react-apollo@2.5.6,
which includes this change to use WatchQueryFetchPolicy:
https://github.com/apollographql/react-apollo/pull/3031/commits/63c2185d286e01a5a5b65e71cd5d80cb315ccc2d