Avoid calling broadcastQueries unless a fetch is happening.
When fetchQuery decides not to make a network request (!shouldFetch), we
can safely assume broadcastQueries will be called later in the fetchQuery
method, after the cache has been updated. Calling broadcastQueries
unconditionally in the middle of the fetchQuery method (before the
fetchRequest call) guarantees an unnecessary extra broadcast when
shouldFetch is false, which can lead to multiple UI renders.