Rename getFieldValue to readField and make it call read functions.
At first I thought it would be risky to allow getFieldValue to call read
functions, because it would open the door to infinite recursion and
expensive chains of read functions.

However, after attempting to write tests that made heavy use of
getFieldValue, I realized that calling read functions is just too useful
to forbid, and, if we crippled getFieldValue in this way, developers would
probably just resort to using cache.readFragment to read data from the
cache, which passes through several more layers of abstraction, and thus
is almost certainly slower than readField.