Make ObservableQuery.variables a getter for this.options.variables.
We can continue to support both this.variables and this.options.variables
for backwards compatibility, but we should not be updating both separately
every time we update variables.

Although this.variables and this.options.variables were almost entirely
synonymous, a brief lag in updating this.variables between setOptions and
setVariables was weirdly important for detecting variable changes in
setVariables, and deciding whether to call fetchQuery.