Bump @apollo/client npm version to 3.0.0-rc.0.
Now that the release candidate phase of testing has begun, we will do our
very best to avoid introducing any new features or breaking API changes,
unless those changes are absolutely necessary to fix bugs.

We are very much aware of a number of outstanding bugs, and we fully
intend to fix all known bugs before the final AC3 release, in addition to
continuing to write and edit the documentation for new AC3 features:
https://github.com/apollographql/apollo-client/milestone/14

If you have been waiting for a signal that the AC3 API is stable/frozen,
this is it. However, if you are not interested in helping to identify and
fix (or at least work around) the remaining issues, then you should wait
for the final release before updating.