Restrict root object identification to ROOT_QUERY. (#6914)
Although "Mutation" and "Subscription" are technically the default names
for the root mutation and subscription types, there should never be any
need to refer to those objects, read from them, or write to them, outside
the normal process of sending and receiving mutations and subscriptions.

If you think those use cases have any value, #6911 will quickly convince
you that the downsides of monopolizing types like Subscription and
Mutation (e.g. assuming they are always singletons) dramatically outweigh
the benefits.

In fact, it's not even possible to request the __typename field for the
root Subscription object, according to the GraphQL specification:
https://spec.graphql.org/June2018/#sec-Single-root-field

In particular, whereas

  query { __typename }

can be used to find out the __typename of the root query object (which
might be something other than "Query" in rare situations), the following
operation is simply illegal, because a subscription must have only a
single root field:

  subscription sub {
    __typename
    newMessage {
      body
      sender
    }
  }

This means the only way to find out that the root Subscription type has a
__typename other than "Subscription" would be through an introspection
query, which is typically disallowed in production.